# docs/swagger.yaml
basePath: /
definitions:
...
  api.createProjectItemRequest:
    properties:
      actual_hours:
        type: number
      assigned_to:
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      name:
        type: string
      priority:
        type: string
      project_id:
        type: string
      status:
        type: string
    required:
    - name
    - project_id
    type: object
...
  domain.ProjectItem:
    properties:
      actual_hours:
        type: number
      assigned_to:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        type: string
      name:
        type: string
      priority:
        type: string
      project_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST in Go with Clean Architecture
  title: Golang API REST
  version: "1.0"
paths:
...
  /v1/project-items:
    get:
      consumes:
      - application/json
      description: Get a list of project items with optional filtering and pagination
      parameters:
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by assigned user ID
        in: query
        name: assigned_to
        type: string
      - description: 'Number of items per page (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Sort order (default: created_at desc)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ProjectItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List project items
      tags:
      - project-items
    post:
      consumes:
      - application/json
      description: Create a new project item
      parameters:
      - description: Project item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createProjectItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ProjectItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create project item
      tags:
      - project-items
  /v1/project-items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a project item (soft delete)
      parameters:
      - description: Project item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete project item
      tags:
      - project-items
    get:
      consumes:
      - application/json
      description: Get a specific project item by its ID
      parameters:
      - description: Project item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProjectItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get project item by ID
      tags:
      - project-items
    put:
      consumes:
      - application/json
      description: Update an existing project item
      parameters:
      - description: Project item ID
        in: path
        name: id
        required: true
        type: string
      - description: Project item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ProjectItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProjectItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update project item
      tags:
      - project-items
  /v1/project-items/project/{projectId}:
    get:
      consumes:
      - application/json
      description: Get all project items for a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ProjectItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get project items by project ID
      tags:
      - project-items
...
swagger: "2.0"